Extract Chat component logic into custom hook
