+++ src/components/Chat/Chat.js
@@ -1,13 +1,11 @@
 // @ts-check
 
-import React, { useState, useEffect, useCallback } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 
-import Dayjs from 'dayjs';
 import { ChatContext, TranslationContext } from '../../context';
-import { Streami18n } from '../../i18n';
 
-import { version } from '../../../package.json';
+import { useChat } from './hooks/useChat';
 
 /**
  * Chat - Wrapper component for Chat. The needs to be placed around any other chat components.
@@ -24,7 +22,7 @@ import { version } from '../../../package.json';
  *
  * @example ../../docs/Chat.md
  * @typedef {import('stream-chat').Channel | undefined} ChannelState
- * @type {React.FC<{ client: import('types').StreamChatReactClient, theme?: string, i18nInstance?: Streami18n, initialNavOpen?: boolean }>}
+ * @type {React.FC<{ client: import('types').StreamChatReactClient, theme?: string, i18nInstance?: import('../../i18n').Streami18n, initialNavOpen?: boolean }>}
  */
 const Chat = ({
   client,
@@ -33,75 +31,15 @@ const Chat = ({
   initialNavOpen = true,
   children,
 }) => {
-  const [translators, setTranslators] = useState(
-    /** @type { Required<import('types').TranslationContextValue>} */ ({
-      t: /** @param {string} key */ (key) => key,
-      tDateTimeParser: (input) => Dayjs(input),
-    }),
-  );
-  const [mutes, setMutes] = useState(
-    /** @type {import('stream-chat').Mute[]} */ ([]),
-  );
-  const [navOpen, setNavOpen] = useState(initialNavOpen);
-  const [channel, setChannel] = useState(
-    /** @type {ChannelState} */ (undefined),
-  );
-
-  const openMobileNav = () => setTimeout(() => setNavOpen(true), 100);
-  const closeMobileNav = () => setNavOpen(false);
-  const clientMutes = client?.user?.mutes;
-
-  useEffect(() => {
-    if (!client) return;
-    const userAgent = client.getUserAgent();
-    if (!userAgent.includes('stream-chat-react')) {
-      // should result in something like:
-      // 'stream-chat-react-2.3.2-stream-chat-javascript-client-browser-2.2.2'
-      client.setUserAgent(`stream-chat-react-${version}-${userAgent}`);
-    }
-  }, [client]);
-
-  useEffect(() => {
-    setMutes(clientMutes || []);
-
-    /** @param {import('stream-chat').Event} e */
-    const handleEvent = (e) => {
-      if (e.type === 'notification.mutes_updated') setMutes(e.me?.mutes || []);
-    };
-    if (client) client.on(handleEvent);
-    return () => client && client.off(handleEvent);
-  }, [client, clientMutes]);
-
-  useEffect(() => {
-    let streami18n;
-    if (i18nInstance instanceof Streami18n) streami18n = i18nInstance;
-    else streami18n = new Streami18n({ language: 'en' });
-
-    streami18n.registerSetLanguageCallback((t) =>
-      setTranslators((prevTranslator) => ({ ...prevTranslator, t })),
-    );
-    streami18n.getTranslators().then((translator) => {
-      if (translator) setTranslators(translator);
-    });
-  }, [i18nInstance]);
-
-  const setActiveChannel = useCallback(
-    /**
-     * @param {ChannelState} activeChannel
-     * @param {{ limit?: number; offset?: number }} [watchers]
-     * @param {React.BaseSyntheticEvent} [e]
-     */
-    async (activeChannel, watchers = {}, e) => {
-      if (e && e.preventDefault) e.preventDefault();
-
-      if (activeChannel && Object.keys(watchers).length) {
-        await activeChannel.query({ watch: true, watchers });
-      }
-      setChannel(activeChannel);
-      closeMobileNav();
-    },
-    [],
-  );
+  const {
+    setActiveChannel,
+    navOpen,
+    mutes,
+    channel,
+    openMobileNav,
+    closeMobileNav,
+    translators,
+  } = useChat({ client, initialNavOpen, i18nInstance });
 
   if (!translators.t) return null;
